requirements.txt
annotated-types==0.7.0
attrs==23.2.0
blinker==1.8.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
dnspython==2.6.1
Flask==3.0.3
h11==0.14.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
mongoengine==0.28.2
numpy==1.26.4
outcome==1.3.0.post0
packaging==24.0
pandas==2.2.2
pydantic_core==2.18.3
pymongo==4.7.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.3
selenium==4.21.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.25.1
trio-websocket==0.11.1
typing_extensions==4.12.0
tzdata==2024.1
urllib3==2.2.1
webdriver-manager==4.0.1
Werkzeug==3.0.4
wsproto==1.2.0
